+++ vice/src/scpu64/scpu64cpu.h
@@ -49,6 +49,6 @@ void maincpu_steal_cycles(void);
 int scpu64_snapshot_write_cpu_state(struct snapshot_module_s *m);
 int scpu64_snapshot_read_cpu_state(struct snapshot_module_s *m);
 extern int scpu64_emulation_mode;
-extern int scpu64_fastmode;
+extern BYTE scpu64_fastmode;
 
 #endif
